Revert "Revert "Merge branch '3-feature/landing_page_v2' into 'master'""
This reverts commit 734c25c9ab744927a54978fe4a90b6d1b4ec8014.

# Conflicts:
#	src/main/java/com/ons/group2/ons_client_project/config/SecurityConfig.java
#	src/main/java/com/ons/group2/ons_client_project/controllers/UserProfileController.java
#	src/main/java/com/ons/group2/ons_client_project/controllers/rest/UserSkillRestController.java
#	src/main/java/com/ons/group2/ons_client_project/model/UserSkill.java
#	src/main/java/com/ons/group2/ons_client_project/model/adaptors/UserSkillAdaptor.java
#	src/main/java/com/ons/group2/ons_client_project/model/dto/skill/NewSkillDto.java
#	src/main/java/com/ons/group2/ons_client_project/service/impl/UserSkillServiceImpl.java
#	src/main/java/com/ons/group2/ons_client_project/web/controllers/pages/PagesController.java
#	src/main/resources/application.properties
#	src/main/resources/schema.sql
#	src/main/resources/templates/profile_page/t_profile_page.html

+++ src/main/java/com/ons/group2/ons_client_project/controllers/UserProfileController.java
@@ -27,7 +27,7 @@ public class UserProfileController {
     }
 
     @GetMapping("/{id}")
-    public String profilePage(@PathVariable("id") Integer id, Model model) {
+    public String profilePage(@PathVariable("id") Long id, Model model) {
 
         Optional<User> user = userService.findById(id);
